+++ api/lib/application/users/index.js
@@ -339,18 +339,6 @@ const register = async function (server) {
     ...adminRoutes,
     {
       method: 'GET',
-      path: '/api/users/me',
-      config: {
-        handler: userController.getCurrentUser,
-        notes: [
-          '- **Cette route est restreinte aux utilisateurs authentifiés**\n' +
-            '- Récupération de l’utilisateur courant\n',
-        ],
-        tags: ['api', 'user'],
-      },
-    },
-    {
-      method: 'GET',
       path: '/api/users/{id}/campaign-participations',
       config: {
         pre: [
