+++ api/src/identity-access-management/application/user/user.route.js
@@ -32,6 +32,18 @@ export const userRoutes = [
     },
   },
   {
+    method: 'GET',
+    path: '/api/users/me',
+    config: {
+      handler: (request, h) => userController.getCurrentUser(request, h),
+      notes: [
+        '- **Cette route est restreinte aux utilisateurs authentifiés**\n' +
+          '- Récupération de l’utilisateur courant\n',
+      ],
+      tags: ['identity-access-management', 'api', 'user'],
+    },
+  },
+  {
     method: 'PATCH',
     path: '/api/users/{id}/password-update',
     config: {
